# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v2/collections/{id}/items": {
      "get": {
        "description": "Retrieve all items (documents) in a collection. Supports cursor-based pagination.",
        "operationId": "ItemsController_find_by_collection_id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order (default: desc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Base64 encoded cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (1-100, default: 25)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of items in the collection"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get collection items",
        "tags": [
          "items"
        ]
      },
      "post": {
        "description": "Add a document to a collection",
        "operationId": "ItemsController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Item data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "document_id": "uuid"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item added to collection successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Add item to collection",
        "tags": [
          "items"
        ]
      }
    },
    "/v2/collections/{id}/items/{itemId}": {
      "delete": {
        "description": "Remove a document from a collection",
        "operationId": "ItemsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "Item UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from collection successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remove item from collection",
        "tags": [
          "items"
        ]
      }
    },
    "/v2/documents": {
      "get": {
        "description": "Retrieve documents with search, filtering, sorting, and cursor-based pagination",
        "operationId": "DocumentsController_find_all",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order (default: desc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort by field (default: created_at)",
            "schema": {
              "enum": [
                "created_at",
                "title",
                "updated_at"
              ],
              "type": "string"
            }
          },
          {
            "name": "created_before",
            "required": false,
            "in": "query",
            "description": "Filter by creation date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "created_after",
            "required": false,
            "in": "query",
            "description": "Filter by creation date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search documents by title (case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folder_id",
            "required": false,
            "in": "query",
            "description": "Filter by folder UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Base64 encoded cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (1-100, default: 25)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/v2/documents/{id}": {
      "get": {
        "description": "Retrieve a specific document's metadata",
        "operationId": "DocumentsController_find_by_id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Document UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Document not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get document by ID",
        "tags": [
          "documents"
        ]
      }
    },
    "/v2/documents/{id}/url": {
      "get": {
        "description": "Generate a signed URL to access the PDF file (expires in 1 hour)",
        "operationId": "DocumentsController_get_document_url",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Document UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "url": "https://...",
                    "expires_in": 3600
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Document not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get document signed URL",
        "tags": [
          "documents"
        ]
      }
    },
    "/v2/folders": {
      "get": {
        "description": "Retrieve folders with pagination. Supports hierarchical structure via parent_id.",
        "operationId": "FoldersController_find_all",
        "parameters": [
          {
            "name": "parent_id",
            "required": false,
            "in": "query",
            "description": "Filter by parent folder UUID (null for root folders)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Base64 encoded cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (1-100, default: 25)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of folders"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all folders",
        "tags": [
          "folders"
        ]
      },
      "post": {
        "description": "Create a new folder. Can be nested by specifying parent_id.",
        "operationId": "FoldersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Folder data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Documents",
                  "parent_id": null
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a folder",
        "tags": [
          "folders"
        ]
      }
    },
    "/v2/folders/{id}": {
      "get": {
        "description": "Retrieve a specific folder",
        "operationId": "FoldersController_find_by_id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Folder UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Folder not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get folder by ID",
        "tags": [
          "folders"
        ]
      },
      "put": {
        "description": "Update an existing folder (full update)",
        "operationId": "FoldersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Folder UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated folder data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "New Name",
                  "parent_id": "uuid"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Folder not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a folder",
        "tags": [
          "folders"
        ]
      },
      "delete": {
        "description": "Remove a folder",
        "operationId": "FoldersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Folder UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Folder not found"
          }
        },
        "security": [
          {
            "api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a folder",
        "tags": [
          "folders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ohara API",
    "description": "REST API for Ohara App developed por EnSer management system with AI agent integration",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Admin API key for server-to-server authentication"
      }
    },
...
  }
}